End of the work on interpreting the input and building the directory tree.

+++ AdventOfCodeDay7/src/main/java/org/nothing/adventofcode/File.java
@@ -1,3 +1,3 @@
 package org.nothing.adventofcode;
 
-public record File (String name, long size){}
+public record File (String name, Double size){}
